move generate_*_report to create_reports
